Replace GeneralFilter with DummyFilter
Code for GeneralFilter is not available and the linker library needs
VS2010 _without_ SP1 from Windows 7.0 SDK to build. So we remove it for
now for ability to build on newer compiler so that we can resume
changes. Restore/replace the filter later, maybe with more advanced
actuator simulations. Released on 2014-12-13.